+++ Sales/ViewModels/ClientViewModels/ClientDisplayViewModel.cs
@@ -37,7 +37,7 @@ namespace Sales.ViewModels.ClientViewModels
             CurrentPage = 1;
             ISFirst = false;
             TotalRecords = clients.Where(w => (w.Name + w.Address + w.Telephone).Contains(_key)).Count();
-            LastPage = (int)Math.Ceiling(Convert.ToDecimal(TotalRecords / 17));
+            LastPage = (int)Math.Ceiling(Convert.ToDecimal((double)TotalRecords / 17));
             if (_lastPage == 0)
                 LastPage = 1;
             if (_lastPage == 1)
