+++ Sales/ViewModels/SpendingViewModels/SpendingReportViewModel.cs
@@ -19,7 +19,7 @@ namespace Sales.ViewModels.SpendingViewModels
             CurrentPage = 1;
             ISFirst = false;
             TotalRecords = spendings.Where(w => (w.Statement).Contains(_key)).Count();
-            LastPage = (int)Math.Ceiling(Convert.ToDecimal(TotalRecords / 17));
+            LastPage = (int)Math.Ceiling(Convert.ToDecimal((double)TotalRecords / 17));
             if (_lastPage == 0)
                 LastPage = 1;
             if (_lastPage == 1)
