+++ Sales/ViewModels/StoreViewModels/CategoryRequiredViewModel.cs
@@ -21,7 +21,7 @@ namespace Sales.ViewModels.StoreViewModels
             CurrentPage = 1;
             ISFirst = false;
             TotalRecords = categories.Where(w => (w.Company + w.Category + w.Stock).Contains(_key)).Count();
-            LastPage = (int)Math.Ceiling(Convert.ToDecimal(TotalRecords / 17));
+            LastPage = (int)Math.Ceiling(Convert.ToDecimal((double)TotalRecords / 17));
             if (_lastPage == 0)
                 LastPage = 1;
             if (_lastPage == 1)
